+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## v1.1.5
 
 - Added more command line options to the executable
+- Added new option: `includeTable`.
 
 ## v1.1.4